requirements.txt
pillow
numpy
pandas 
tensorflow
keras
scikit-learn
